move calculators to separate folder
